server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.5",
    "aiosqlite>=0.22.1",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
]
